Move user test code to separate file (and color printed code)
